Badly solves memory leak. See commit message
By curiosity i went to check activity monitor on my mac and i found that my program was using 59gb of RAM!! I need to solve this.

I was able to solve the memory leak by tracing what Idid recently. Ifound out that animation was the root cause of this. When i introduce the animated flag to the SpriteComponent. Deleting the whole block code of animation solved the memory leak.

Once i figured that the problem was there i started looking at the code to see if i could spot it. Eventually I spotted it. I saw that I was calling `setTexture` which then called `TextureManager::LoadTexture` every frame on the game loop -- the update method. Deleting that line also solved the memory leak. So the problem was there.

My first thought was to create a vector of textures for each of my animations vector so that I could reuse them. But that would make me write more code. So i went in a different path.

I noticed that `TextureManager::LoadTexture` was creating a pointer that was not used anywhere and it was also not cleaned up: `SDL_Surface *tempSurface = IMG_Load(fileName);`. I read on a book that everytime we created a pointer we should clean up the pointer, or else we'd have a dangling pointer, and that was exactly what I had here.

Reading online I found that we could free the memory surface `SDL_FreeSurface(tempSurface);` doing this already solved part of the problem, the leak as much smaller and memory was growing much slower.

I also knew that we had to clean the texture because we previously had clean it up on destructors so on the `setTexture` I also called `SDL_DestroyTexture(this->texture);` before loading the next texture.

That solved the memory leak entirely! Nice.

Now my first thought was much smarter though. This was just 2 lines of code to solve the problem, but I now had another problem. I went to Claude, to clarify if my code was correct and if it made sense.

He reviewed my code and said that `SDL_FreeSurface` was the right thing, because we no longer use it. We first load the surface from disk to cpu, and then the texture is sent to the GPU.

The problem here was that I was doing that in everyframe, which he said was, in his own words: **This is a performance disaster.**

```
Every frame:

- SDL_DestroyTexture - frees GPU memory
- IMG_Load - reads file from disk (SLOW)
- SDL_CreateTextureFromSurface - uploads to GPU
- SDL_FreeSurface - frees RAM

This is incredibly slow. Disk I/O every frame will kill your framerate.
What you should do instead:
- Load all animation textures ONCE (in constructor or init)
- Delete all textures in the destructor
```

Now in my next commit I'll do the proper way.